Using numbers to specify import order in Kibana.
In dashboard-importer/kibana-config we have a list of json files that
must be applied in a specified order. I suggest to use numbers to group
the various files instead of couple of letters as before.
This way the elastic search importer will load all the 1_ before the 2_
and so on. 3_ are all the visualizations that can be imported with
random order.

Change-Id: Ia968506b9628b86505323f7cb3ad6a9f9c815951