Allow the installation of analytics-etl as plugin
Introduce an SSH command to start the extraction of Gerrit Analytics
and re-package the fat-jar as Gerrit plugin.
To run the Spark job inside Gerrit, requires to have the Spark
libraries linked into the /lib directory beforehand.

Keep the ability to still run the Job in a Spark Cluster by keeping
the main entry point and the Spark dependencies as provided.

Bug: Issue 9021
Change-Id: I66a051250242b1534400d6baacad0cdaaf177266