Update patch set 5
Patch Set 5:

> Patch Set 5: Code-Review-1
> 
> (4 comments)
> 
> Hey Tiago,
> I think I am a bit lost on what the purpose of this script is, because there are some bits I don't understand. perhaps I am just lacking some context, in which case, apologies :)

Hi Antonio,

The purpose of this script is to setup local dev environment where you can then test the ETL changes end2end/integration tests.

It starts a gerrit instance with a list of cloned repositories and starts the ES with kibana. 
I found this useful when I wanted to test a few new dashbords/metrics

This script is not supposed to run the ETL application when it starts.

Patch-set: 5
Reviewer: Gerrit User 1065251 <1065251@173816e5-2b9a-37c3-8a2e-48639d4f1153>