Provide Spark job to process AuditLogs events
This change provides a new ETL job that allows the processing of auditLog
events (typically generated by the audit-sl4j plugin).
Events are normalized, aggregated and then persisted to an elasticsearch
index. An indended use of this is to then expose this information through
visualization dashboards, such as kibana.

This work also refactors some general purpose code in order to be shared
by all the ETLs, such as connection to gerrit APIand time operations.

Change-Id: I9d1fd9ae569de3b5c0d64cb625356fad3a62fd7a