Remove Change processing from events stream
This feature is not currently used and it is not working as it is.
It was a POC to check the possibility of combining batch and stream events processing.
The code is getting in the way of new features development, hence
the decision of removing it, until the necessity of stream processing will arise again.

Feature: Issue 10211
Change-Id: I4a3e993b80ae0c43fa5248ec2acfe10da9b6ed5c