Enrich user info in Auditlogs ETL
Allow enrichment of user information. This change on allows "user type" to be defined, but it can
be easily extended to any other useful information.

User data is passed via a CSV file with the following format:
id,type
123,bot
456,human

"User type" is defaulted to "human" if not present in this file.

A possible use case for this feature is the distinction between "bot" and "human" users.

Feature: Issue 10229
Change-Id: I042244154e668f5137ff97053f1cc2ad88342acd