Leverage bot-like-regexps parameter of analytics plugin
This commit makes use of the bot-like-regexps GET parameter of the
analytics plugin, which allows to pass a comma-separated list of regular
expressions to use in deciding whether a particular commit looks like a
bot-like commit.

This change also index the boolean is_bot_like field in elasticsearch
so that it can be used to build analytics on.

Change-Id: I12620ab37eb0866210ec1a690c827b459481d36e