Store project revision in ES document
The ETL correctly checks out the version specified
in the manifest for each project, but it doesn't store
the project revision in the Elasticsearch document.

This change ensures that the project revision is captured
and stored in the ES document.

This field is only populated when using import via manifest.

For example, given the following manifest snippet:
```
<project name="p1" revision="branch1"/>
<project name="p2" revision="branch2"/>
```

The revisions branch1 and branch2 will now be
stored in the ES document for their respective projects.

Change-Id: I0e394006b4f008d140859af3cc24874442672e9e